Support for custom types of pagelets defined by a project to appear in pages.
Custom pagelet types inherit from PageletBase and define their own
render(). They are registered in settings with
PAGELET_ADDITIONAL_PAGELET_TYPES, a list of app.model strings.

Custom types have an order and area of the page to render in, but they
do not have a slug, unless they decide to define one. If they appear
with an edit link, the link will redirect to the admin.